feat: 2023 day 23a
What happened to 23b? I solved it accidentally. I started a bruteforce search on
part b, but as it wasn't finishing I started working on a better solution. I
forgot that I left it running in the background and after a few hours it solved
the problem. I will commit a better solution once I finish it.